requirements.txt
# Core Discord bot dependencies
discord.py>=2.0.0
python-dotenv>=1.0.0

# Ban stats bot dependencies
matplotlib>=3.7.0

# Random colors bot dependencies  
httpx>=0.24.0
webcolors>=1.13.0

# Summary bot dependencies
mistralai>=1.0.0

# Development/deployment dependencies
pm2>=5.3.0